Fix: 统一 HTTP 客户端并启用 reqwest socks 特性
- Cargo: 启用 reqwest socks，移除 blocking
- 新增 http_client 模块，集中构建带代理的 reqwest Client
- VersionService 改为使用统一 Client，替换 reqwest::get
- main.rs 的构建函数委托到统一实现

Fixes Duckcoding-dev/duckcoding#18
