+++ src-tauri/Cargo.toml
@@ -18,7 +18,7 @@ serde_json = "1"
 dirs = "6"
 toml = "0.9"
 toml_edit = "0.23"
-reqwest = { version = "0.12", features = ["json", "blocking"] }
+reqwest = { version = "0.12", features = ["json", "socks"] }
 tokio = { version = "1", features = ["full"] }
 urlencoding = "2.1"
 regex = "1"
